+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @module-federation/runtime
 
+## 0.0.7
+
+### Patch Changes
+
+- 7df24df: only both version and name matched instance can be re-use
+- 7df24df: feat: add initContainer and beforeInitContainer hook
+  - @module-federation/sdk@0.0.7
+
+## 0.0.6
+
+### Patch Changes
+
+- b505deb: fix: rename usePlugins to prevent swc react-refresh from throwing errors when replacing variables
+  - @module-federation/sdk@0.0.6
+
 ## 0.0.5
 
 ### Patch Changes